Explicitly cast the return of `luaL_checknumber` to int.
  This avoids an implicit conversion

+++ bindings.c
@@ -259,7 +259,7 @@ int lua_above(lua_State * L)
  */
 int lua_activate_workspace(lua_State * L)
 {
-    int number = luaL_checknumber(L, 1);
+    int number = (int)luaL_checknumber(L, 1);
 
     /**
      * Get the count of workspaces.
@@ -531,8 +531,8 @@ int lua_size(lua_State * L)
      */
     if (top > 0)
     {
-        int h = luaL_checknumber(L, 1);
-        int w = luaL_checknumber(L, 2);
+        int h = (int)luaL_checknumber(L, 1);
+        int w = (int)luaL_checknumber(L, 2);
 
 
         char *x = g_strdup_printf("size(%d,%d);", h, w);
@@ -739,7 +739,7 @@ int lua_workspace(lua_State * L)
      */
     if (top > 0)
     {
-        int number = luaL_checknumber(L, 1);
+        int number = (int)luaL_checknumber(L, 1);
 
 
         /**
@@ -798,7 +798,7 @@ int lua_workspaces(lua_State * L)
      */
     if (top > 0)
     {
-        int newcount = luaL_checknumber(L, 1);
+        int newcount = (int)luaL_checknumber(L, 1);
         wnck_screen_change_workspace_count(screen, newcount);
     }
 
@@ -820,8 +820,8 @@ int lua_xy(lua_State * L)
      */
     if (top > 0)
     {
-        int newx = luaL_checknumber(L, 1);
-        int newy = luaL_checknumber(L, 2);
+        int newx = (int)luaL_checknumber(L, 1);
+        int newy = (int)luaL_checknumber(L, 2);
 
 
         char *x = g_strdup_printf("xy(%d,%d);", newx, newy);
